Split tests into multiple files
This commit:

- Separates unit tests into multiple files
- Separates feature tests into multiple files
- Gets feature tests working
