+++ darts/models.py
@@ -6,9 +6,3 @@ class Document(models.Model):
     file = models.BinaryField()
     body = models.TextField()
     search_text = SearchVectorField()
-
-    @classmethod
-    def create(cls, title):
-        book = cls(title=title)
-        book.save()
-        return book
